[PDF Rspecs] refactoring
